Physical Access Control System(PACS) AIS/requirements.txt
django==4.2.7
django-stubs==4.2.6
djangorestframework==3.14.0
drf-yasg==1.21.7
PyMySql==1.1.0
mysqlclient==2.2.0
gunicorn==21.2.0